requirements-dev.txt
flask==1.1.2
flask-restx==0.2.0
Flask-SQLAlchemy==2.4.4
psycopg2-binary==2.8.6
requests==2.25.1
Flask-SQLAlchemy==2.4.4
simplejson==3.17.2